added homefeed remaining changes
