+++ ChatSX/screens/HomeFeed/Tabs/DMFeed/index.tsx
@@ -155,13 +155,13 @@ const DMFeed = ({navigation}: Props) => {
     }
   };
 
-  useEffect(() => {
-    if (isFocused) {
-      getExistingData();
-      if (!user?.sdkClientInfo?.community) return;
-      paginatedSyncAPI(INITIAL_SYNC_PAGE, user?.sdkClientInfo?.community);
-    }
-  }, [isFocused, user]);
+  // useEffect(() => {
+  //   if (isFocused) {
+  //     getExistingData();
+  //     if (!user?.sdkClientInfo?.community) return;
+  //     paginatedSyncAPI(INITIAL_SYNC_PAGE, user?.sdkClientInfo?.community);
+  //   }
+  // }, [isFocused, user]);
 
   const listener = async () => {
     const chatroomObservable = new Observable(observer => {
@@ -203,9 +203,9 @@ const DMFeed = ({navigation}: Props) => {
     };
   };
 
-  useEffect(() => {
-    listener();
-  }, [isFocused]);
+  // useEffect(() => {
+  //   listener();
+  // }, [isFocused]);
 
   // useLayoutEffect(() => {
   //   fetchData();
@@ -267,15 +267,15 @@ const DMFeed = ({navigation}: Props) => {
     ) : null;
   };
 
-  useEffect(() => {
-    const query = ref(db, `/community/${community?.id}`);
-    return onValue(query, snapshot => {
-      if (snapshot.exists()) {
-        if (!user?.sdkClientInfo?.community) return;
-        paginatedSyncAPI(INITIAL_SYNC_PAGE, user?.sdkClientInfo?.community);
-      }
-    });
-  }, [user]);
+  // useEffect(() => {
+  //   const query = ref(db, `/community/${community?.id}`);
+  //   return onValue(query, snapshot => {
+  //     if (snapshot.exists()) {
+  //       if (!user?.sdkClientInfo?.community) return;
+  //       paginatedSyncAPI(INITIAL_SYNC_PAGE, user?.sdkClientInfo?.community);
+  //     }
+  //   });
+  // }, [user]);
 
   return (
     <View style={styles.page}>
